add csrf protection

+++ src/Entrypoint/Controller/Keyforge/Deck/UpdateOwnership/UpdateDeckOwnershipController.php
@@ -39,10 +39,18 @@ final class UpdateDeckOwnershipController extends Controller
         }
 
         if ($request->getMethod() === Request::METHOD_POST) {
+            if (false === $this->isCsrfTokenValid('keyforge_ownership_update_add', $request->get('_csrf_token'))) {
+                throw new \Exception('Invalid CSRF token');
+            }
+
             $this->deckRepository->addOwner(Uuid::from($id), $user->id());
         }
 
         if ($request->getMethod() === Request::METHOD_DELETE) {
+            if (false === $this->isCsrfTokenValid('keyforge_ownership_update_remove', $request->get('_csrf_token'))) {
+                throw new \Exception('Invalid CSRF token');
+            }
+
             $this->deckRepository->removeOwner(Uuid::from($id), $user->id());
         }
